small restructuring
